Change sorting state management to a context
